requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.17
langchain-pinecone==0.1.1
pinecone-client==4.1.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pypdf==4.3.1
python-docx==1.1.2
docx2txt==0.8
markdownify==0.11.6
langchain-text-splitters==0.2.2
tiktoken==0.7.0
tenacity==9.0.0
httpx==0.27.0
pandas==2.2.3
numpy==1.26.4
rich==13.7.1
typer==0.12.4
uvicorn==0.30.5
fastapi==0.112.0
langgraph==0.2.21
pytest==8.3.2
